requirements.txt
tensorflow_gpu==2.0.0
opencv_python==4.1.2.30
numpy==1.19.0
streamlit==0.63.1
Pillow==7.2.0
tensorflow==2.3.0
